+++ frontend/src/metabase-lib/lib/Question.js
@@ -1,5 +1,5 @@
 import _ from "underscore";
-import { chain, assoc, dissoc, assocIn, getIn } from "icepick";
+import { assoc, assocIn, chain, dissoc, getIn } from "icepick";
 
 // NOTE: the order of these matters due to circular dependency issues
 import StructuredQuery, {
@@ -28,11 +28,10 @@ import { isStandard } from "metabase/lib/query/filter";
 import { memoize, sortObject } from "metabase-lib/lib/utils";
 
 // TODO: remove these dependencies
-import * as Card_DEPRECATED from "metabase/lib/card";
 import * as Urls from "metabase/lib/urls";
 import {
-  findColumnSettingIndexForColumn,
   findColumnIndexForColumnSetting,
+  findColumnSettingIndexForColumn,
   syncTableColumnsToQuery,
 } from "metabase/lib/dataset";
 import {
@@ -40,19 +39,19 @@ import {
   isTransientId,
 } from "metabase/meta/Card";
 import {
-  parameterToMBQLFilter,
   normalizeParameterValue,
+  parameterToMBQLFilter,
 } from "metabase/meta/Parameter";
 import {
   aggregate,
   breakout,
+  distribution,
+  drillUnderlyingRecords,
   filter,
   pivot,
-  distribution,
   toUnderlyingRecords,
-  drillUnderlyingRecords,
 } from "metabase/modes/lib/actions";
-import { MetabaseApi, CardApi, maybeUsePivotEndpoint } from "metabase/services";
+import { CardApi, maybeUsePivotEndpoint, MetabaseApi } from "metabase/services";
 import Questions from "metabase/entities/questions";
 
 import type {
@@ -60,8 +59,8 @@ import type {
   ParameterValues,
 } from "metabase-types/types/Parameter";
 import type {
-  DatasetQuery,
   Card as CardObject,
+  DatasetQuery,
   VisualizationSettings,
 } from "metabase-types/types/Card";
 import type { Dataset, Value } from "metabase-types/types/Dataset";
@@ -74,6 +73,7 @@ import {
   ALERT_TYPE_ROWS,
   ALERT_TYPE_TIMESERIES_GOAL,
 } from "metabase-lib/lib/Alert";
+import { utf8_to_b64url } from "metabase/lib/encoding";
 
 type QuestionUpdateFn = (q: Question) => ?Promise<void>;
 
@@ -849,15 +849,13 @@ export default class Question {
     let cellQuery = "";
     if (filters.length > 0) {
       const mbqlFilter = filters.length > 1 ? ["and", ...filters] : filters[0];
-      cellQuery = `/cell/${Card_DEPRECATED.utf8_to_b64url(
-        JSON.stringify(mbqlFilter),
-      )}`;
+      cellQuery = `/cell/${utf8_to_b64url(JSON.stringify(mbqlFilter))}`;
     }
     const questionId = this.id();
     if (questionId != null && !isTransientId(questionId)) {
       return `/auto/dashboard/question/${questionId}${cellQuery}`;
     } else {
-      const adHocQuery = Card_DEPRECATED.utf8_to_b64url(
+      const adHocQuery = utf8_to_b64url(
         JSON.stringify(this.card().dataset_query),
       );
       return `/auto/dashboard/adhoc/${adHocQuery}${cellQuery}`;
@@ -868,9 +866,7 @@ export default class Question {
     let cellQuery = "";
     if (filters.length > 0) {
       const mbqlFilter = filters.length > 1 ? ["and", ...filters] : filters[0];
-      cellQuery = `/cell/${Card_DEPRECATED.utf8_to_b64url(
-        JSON.stringify(mbqlFilter),
-      )}`;
+      cellQuery = `/cell/${utf8_to_b64url(JSON.stringify(mbqlFilter))}`;
     }
     const questionId = this.id();
     const query = this.query();
@@ -880,7 +876,7 @@ export default class Question {
         if (questionId != null && !isTransientId(questionId)) {
           return `/auto/dashboard/question/${questionId}${cellQuery}/compare/table/${tableId}`;
         } else {
-          const adHocQuery = Card_DEPRECATED.utf8_to_b64url(
+          const adHocQuery = utf8_to_b64url(
             JSON.stringify(this.card().dataset_query),
           );
           return `/auto/dashboard/adhoc/${adHocQuery}${cellQuery}/compare/table/${tableId}`;
@@ -1083,7 +1079,7 @@ export default class Question {
         : {}),
     };
 
-    return Card_DEPRECATED.utf8_to_b64url(JSON.stringify(sortObject(cardCopy)));
+    return utf8_to_b64url(JSON.stringify(sortObject(cardCopy)));
   }
 
   convertParametersToFilters() {
