+++ frontend/src/metabase/lib/card.js
@@ -4,6 +4,7 @@ import Utils from "metabase/lib/utils";
 import * as Urls from "metabase/lib/urls";
 
 import { CardApi } from "metabase/services";
+import { b64url_to_utf8, utf8_to_b64url } from "metabase/lib/encoding";
 
 export function createCard(name = null) {
   return {
@@ -92,25 +93,6 @@ export function serializeCardForUrl(card) {
 export function deserializeCardFromUrl(serialized) {
   serialized = serialized.replace(/^#/, "");
   return JSON.parse(b64url_to_utf8(serialized));
-}
-
-// escaping before base64 encoding is necessary for non-ASCII characters
-// https://developer.mozilla.org/en-US/docs/Web/API/WindowBase64/btoa
-export function utf8_to_b64(str) {
-  return window.btoa(unescape(encodeURIComponent(str)));
-}
-export function b64_to_utf8(b64) {
-  return decodeURIComponent(escape(window.atob(b64)));
-}
-
-// for "URL safe" base64, replace "+" with "-" and "/" with "_" as per RFC 4648
-export function utf8_to_b64url(str) {
-  return utf8_to_b64(str)
-    .replace(/\+/g, "-")
-    .replace(/\//g, "_");
-}
-export function b64url_to_utf8(b64url) {
-  return b64_to_utf8(b64url.replace(/-/g, "+").replace(/_/g, "/"));
 }
 
 export function urlForCardState(state, dirty) {
